remove nonexistent about page from main target and tests
